+++ fastapi_cache/types.py
@@ -1,3 +1,4 @@
+import abc
 from typing import Any, Awaitable, Callable, Dict, Optional, Tuple, Union
 
 from starlette.requests import Request
@@ -20,3 +21,21 @@ class KeyBuilder(Protocol):
         kwargs: Dict[str, Any],
     ) -> Union[Awaitable[str], str]:
         ...
+
+
+class Backend(abc.ABC):
+    @abc.abstractmethod
+    async def get_with_ttl(self, key: str) -> Tuple[int, Optional[bytes]]:
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    async def get(self, key: str) -> Optional[bytes]:
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    async def set(self, key: str, value: bytes, expire: Optional[int] = None) -> None:
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    async def clear(self, namespace: Optional[str] = None, key: Optional[str] = None) -> int:
+        raise NotImplementedError
